refactor: move templates to own lib
