+++ apps/web/pages/_document.tsx
@@ -10,7 +10,7 @@ export default class MyDocument extends Document {
   render() {
     return (
       <Html>
-        <Head></Head>
+        <Head />
         <body>
           <Main />
           <NextScript />
