+++ apps/web/pages/admin/index.tsx
@@ -1,15 +1,11 @@
 import { withPageAuthRequired } from '@auth0/nextjs-auth0'
 import { NextPageTemplate } from '@codelab/frontend/abstract/props'
+import { MainDashboardTemplate } from '@codelab/frontend/view/templates'
 import { PageHeader } from 'antd'
-import { MainDashboardTemplate } from 'apps/web/src/templates/MainDashboardTemplate'
 import React from 'react'
 
 const AdminPage: NextPageTemplate<'dashboard'> = () => {
-  return (
-    <>
-      <PageHeader ghost={false} title="Admin" />
-    </>
-  )
+  return <PageHeader ghost={false} title="Admin" />
 }
 
 AdminPage.Template = MainDashboardTemplate
