+++ apps/web/pages/apps/[appId]/pages/[pageId]/builder.tsx
@@ -1,8 +1,14 @@
 import { withPageAuthRequired } from '@auth0/nextjs-auth0'
+import { NextPageTemplate } from '@codelab/frontend/abstract/props'
 import {
   Builder,
   defaultRenderContext,
 } from '@codelab/frontend/modules/builder'
+import {
+  MainPanePageDetail,
+  MetaPanePageDetail,
+  PageDetailHeader,
+} from '@codelab/frontend/modules/page'
 import {
   PageContext,
   RenderProvider,
@@ -10,10 +16,6 @@ import {
 } from '@codelab/frontend/presenter/container'
 import { Empty } from 'antd'
 import React, { useContext } from 'react'
-import { MainPanePageDetail } from '../../../../../../../libs/frontend/modules/page/src/main-pane'
-import { MetaPanePageDetail } from '../../../../../../../libs/frontend/modules/page/src/meta-pane'
-import { PageDetailHeader } from '../../../../../src/sections/header'
-import { NextPageTemplate } from '../../../../../src/templates/Layout.interface'
 import { PageBuilderTemplate } from '../../../../../src/templates/PageBuilderTemplate'
 
 const PageBuilder: NextPageTemplate<'builder'> = () => {
