+++ apps/web/pages/apps/index.tsx
@@ -1,4 +1,5 @@
 import { withPageAuthRequired } from '@auth0/nextjs-auth0'
+import { NextPageTemplate } from '@codelab/frontend/abstract/props'
 import {
   CreateAppButton,
   CreateAppModal,
@@ -10,7 +11,6 @@ import { SignOutUserButton } from '@codelab/frontend/modules/user'
 import { padding } from '@codelab/frontend/style'
 import { PageHeader } from 'antd'
 import React from 'react'
-import { NextPageTemplate } from '../../src/templates/Layout.interface'
 import { MainDashboardTemplate } from '../../src/templates/MainDashboardTemplate'
 
 const AppsPage: NextPageTemplate<'dashboard'> = () => {
