+++ apps/web/pages/index.tsx
@@ -1,9 +1,9 @@
+import { NextPageTemplate } from '@codelab/frontend/abstract/props'
 import React from 'react'
-import { HomeClients } from '../src/sections/home/Home-clients'
-import { HomeFeatures } from '../src/sections/home/Home-features'
-import { HomeJumbo } from '../src/sections/home/Home-jumbo'
+import { HomeClients } from '../src/home/Home-clients'
+import { HomeFeatures } from '../src/home/Home-features'
+import { HomeJumbo } from '../src/home/Home-jumbo'
 import { HomeTemplate } from '../src/templates/HomeTemplate'
-import { NextPageTemplate } from '../src/templates/Layout.interface'
 
 const HomePage: NextPageTemplate = () => {
   return (
