+++ apps/web/pages/tags/index.tsx
@@ -1,4 +1,5 @@
 import { withPageAuthRequired } from '@auth0/nextjs-auth0'
+import { NextPageTemplate } from '@codelab/frontend/abstract/props'
 import {
   CreateTagButton,
   CreateTagModal,
@@ -6,10 +7,9 @@ import {
   GetTagsTree,
   UpdateTagModal,
 } from '@codelab/frontend/modules/tag'
+import { MainDashboardTemplate } from '@codelab/frontend/view/templates'
 import { PageHeader } from 'antd'
 import React from 'react'
-import { NextPageTemplate } from '../../src/templates/Layout.interface'
-import { MainDashboardTemplate } from '../../src/templates/MainDashboardTemplate'
 
 const TagPage: NextPageTemplate<'dashboard'> = () => {
   const pageHeaderButtons = [<CreateTagButton key={0} />]
@@ -26,11 +26,7 @@ const TagPage: NextPageTemplate<'dashboard'> = () => {
 }
 
 TagPage.Template = MainDashboardTemplate
-TagPage.MainPane = () => (
-  <>
-    <GetTagsTree />
-  </>
-)
+TagPage.MainPane = () => <GetTagsTree />
 
 export const getServerSideProps = withPageAuthRequired()
