+++ apps/web/pages/types/index.tsx
@@ -1,4 +1,5 @@
 import { withPageAuthRequired } from '@auth0/nextjs-auth0'
+import { NextPageTemplate } from '@codelab/frontend/abstract/props'
 import {
   CreateTypeButton,
   CreateTypeModal,
@@ -7,9 +8,8 @@ import {
   UpdateTypeModal,
 } from '@codelab/frontend/modules/type'
 import { padding } from '@codelab/frontend/style'
+import { MainDashboardTemplate } from '@codelab/frontend/view/templates'
 import { PageHeader } from 'antd'
-import { NextPageTemplate } from 'apps/web/src/templates/Layout.interface'
-import { MainDashboardTemplate } from 'apps/web/src/templates/MainDashboardTemplate'
 import React from 'react'
 
 const TypesPage: NextPageTemplate<'dashboard'> = () => {
