+++ apps/web/src/home/index.ts
@@ -1,0 +1,1 @@
+export * from '../templates/HomeTemplate'
