--- apps/web/src/sections/header/index.ts
@@ -1,1 +1,0 @@
-export * from './PageDetailHeader'
