--- apps/web/src/sections/home/index.ts
@@ -1,1 +1,0 @@
-export * from '../../templates/HomeTemplate'
