--- apps/web/src/sections/panes/index.ts
@@ -1,5 +1,0 @@
-export * from './component'
-export * from './lambda'
-export * from './library'
-export * from './page'
-export * from './paneTemplates'
