--- apps/web/src/sections/panes/lambda/index.ts
@@ -1,1 +1,0 @@
-export * from './MainPaneLambda'
