--- apps/web/src/sections/panes/library/index.ts
@@ -1,1 +1,0 @@
-export * from './MainPaneLibrary'
