--- apps/web/src/sections/panes/page/index.ts
@@ -1,2 +1,0 @@
-export * from './mainPane'
-export * from './metaPane'
