+++ apps/web/src/templates/HomeTemplate.tsx
@@ -1,8 +1,8 @@
+import { LayoutComponent } from '@codelab/frontend/abstract/props'
 import { contentStyle } from '@codelab/frontend/style'
 import { Layout } from 'antd'
 import React from 'react'
-import { HomeMenuHeader } from '../sections/home/Home-menu--header'
-import { LayoutComponent } from './Layout.interface'
+import { HomeMenuHeader } from '../home/Home-menu--header'
 
 const { Content, Header, Footer } = Layout
