+++ libs/frontend/modules/app/src/index.ts
@@ -1,4 +1,4 @@
-export * from './createApp'
-export * from './deleteApp'
-export * from './getApps'
-export * from './updateApp'
+export * from './use-cases/create-app'
+export * from './use-cases/delete-app'
+export * from './use-cases/get-apps'
+export * from './use-cases/update-app'
