+++ libs/frontend/modules/builder/src/index.ts
@@ -5,6 +5,7 @@ export * from './EditorProvider'
 export * from './main-pane'
 export * from './MetaPaneBuilderTemplate'
 export * from './overlay-toolbar'
+export * from './PageBuilderTemplate'
 export * from './renderer'
 export * from './useBuilder'
 export * from './useBuilderHandlers'
