+++ libs/frontend/modules/component/src/ComponentDetailTemplate.tsx
@@ -1,7 +1,7 @@
+import { LayoutComponent } from '@codelab/frontend/abstract/props'
 import { EditorProvider } from '@codelab/frontend/modules/builder'
 import React from 'react'
 import { DashboardLayout } from '../layout/DashboardLayout'
-import { LayoutComponent } from './Layout.interface'
 
 export const ComponentDetailTemplate: LayoutComponent<'builder'> = (props) => {
   const { children, MainPane, MetaPane } = props
