+++ libs/frontend/modules/component/src/index.ts
@@ -1,3 +1,8 @@
+export * from './ComponentDetailTemplate'
 export * from './main-pane'
 export * from './meta-pane/MetaPaneComponentDetail'
-export * from './use-cases'
+export * from './use-cases/component-element'
+export * from './use-cases/create-component'
+export * from './use-cases/delete-component'
+export * from './use-cases/get-components'
+export * from './use-cases/update-component'
