--- libs/frontend/modules/component/src/use-cases/index.ts
@@ -1,4 +1,0 @@
-export * from './component-element'
-export * from './create-component'
-export * from './delete-component'
-export * from './update-component'
