+++ libs/frontend/modules/lambda/src/MainPaneLambda.tsx
@@ -1,11 +1,11 @@
+import { MainPaneTemplate } from '@codelab/frontend/view/templates'
+import React from 'react'
 import {
   CreateLambdaButton,
   CreateLambdaModal,
-  GetLambdasTable,
-  UpdateLambdaModal,
-} from '@codelab/frontend/modules/lambda'
-import { MainPaneTemplate } from '@codelab/frontend/view/templates'
-import React from 'react'
+} from './use-cases/create-lambda'
+import { GetLambdasTable } from './use-cases/get-lambdas'
+import { UpdateLambdaModal } from './use-cases/update-lambda'
 
 export const MainPaneLambda = () => {
   return (
