+++ libs/frontend/modules/library/src/MainPaneLibrary.tsx
@@ -1,12 +1,12 @@
+import { MainPaneTemplate } from '@codelab/frontend/view/templates'
+import React from 'react'
+import { LibraryExplorerTree } from './library-explorer'
 import {
   CreateLibraryButton,
   CreateLibraryModal,
-  DeleteLibraryModal,
-  LibraryExplorerTree,
-  UpdateLibraryModal,
-} from '@codelab/frontend/modules/library'
-import React from 'react'
-import { MainPaneTemplate } from '../paneTemplates'
+} from './use-cases/create-library'
+import { DeleteLibraryModal } from './use-cases/delete-library'
+import { UpdateLibraryModal } from './use-cases/update-library'
 
 export const MainPaneLibrary = () => {
   return (
