+++ libs/frontend/modules/library/src/index.ts
@@ -1,5 +1,7 @@
 export * from './library-explorer'
 export * from './librarySelectionState'
+export * from './LibraryTemplate'
+export * from './MainPaneLibrary'
 export * from './use-cases/create-library'
 export * from './use-cases/delete-library'
 export * from './use-cases/update-library'
