+++ libs/frontend/modules/page/src/PageDetailHeader.tsx
