+++ libs/frontend/view/templates/DashboardSidebarNavigation.tsx
