+++ libs/frontend/view/templates/index.ts
@@ -1,1 +1,4 @@
+export * from './DashboardLayout'
+export * from './DashboardSidebarNavigation'
+export * from './MainDashboardTemplate'
 export * from './MainPaneTemplate'
